修改：     Rangenet++/train/tasks/semantic/dataset/kitti/parser.py 	新文件：   Rangenet++/train/tasks/semantic/log/arch_cfg.yaml 	新文件：   Rangenet++/train/tasks/semantic/log/data_cfg.yaml 	新文件：   Rangenet++/train/tasks/semantic/log/tb/events.out.tfevents.1709444375.cyj-ROG-Strix.76187.0.v2 	修改：     Rangenet++/train/tasks/semantic/modules/mseEval.py 	修改：     Rangenet++/train/tasks/semantic/modules/trainer.py 	修改：     Rangenet++/train/tasks/semantic/modules/user.py 	新文件：   datasets/2011_09_26/calib_cam_to_cam.txt 	新文件：   datasets/2011_09_26/calib_imu_to_velo.txt 	新文件：   datasets/2011_09_26/calib_velo_to_cam.txt 	新文件：   datasets/2011_09_30/calib_cam_to_cam.txt 	新文件：   datasets/2011_09_30/calib_imu_to_velo.txt 	新文件：   datasets/2011_09_30/calib_velo_to_cam.txt 	新文件：   datasets/2011_10_03/calib_cam_to_cam.txt 	新文件：   datasets/2011_10_03/calib_imu_to_velo.txt 	新文件：   datasets/2011_10_03/calib_velo_to_cam.txt 	修改：     unit_operations/2.coor_transformer.ipynb 	修改：     unit_operations/3.ICP.ipynb 	修改：     unit_operations/7.ICP_label_maker.ipynb 	新文件：   unit_operations/temp.ipynb
尚未暂存以备提交的变更：
	删除：     unit_operations/temp.ipynb

+++ Rangenet++/train/tasks/semantic/dataset/kitti/parser.py
@@ -150,7 +150,7 @@ class SemanticKitti(Dataset):
     unproj_remissions = torch.full([self.max_points], -1.0, dtype=torch.float)
     unproj_remissions[:unproj_n_points] = torch.from_numpy(scan.remissions)
     if self.gt:
-      unproj_labels = torch.full([self.max_points], -1.0, dtype=torch.int32)
+      unproj_labels = torch.full([self.max_points], -1.0, dtype=torch.float32)
       unproj_labels[:unproj_n_points] = torch.from_numpy(scan.sem_label)
     else:
       unproj_labels = []
